gameManager userId ok

+++ backend/src/models/GameManager.js
@@ -9,11 +9,11 @@ class GameManager extends AbstractManager {
 
   // The C of CRUD - Create operation
 
-  async create(game) {
+  async create(userId, game) {
     // Execute the SQL INSERT query to add a new game to the "game" table
     const [result] = await this.database.query(
-      `insert into ${this.table} (title, image, year, console, genre_id) values (?, ? , ? , ? , ?)`,
-      [game.title, game.image, game.year, game.console, game.genre_id]
+      `insert into ${this.table} (title, image, year, console, genre_id, user_id) values (?, ? , ? , ? , ?, ?)`,
+      [game.title, game.image, game.year, game.console, game.genre_id, userId]
     );
 
     // Return the ID of the newly inserted game
@@ -22,26 +22,26 @@ class GameManager extends AbstractManager {
 
   // The Rs of CRUD - Read operations
 
-  async read(id) {
+  async read(userId, id) {
     // Execute the SQL SELECT query to retrieve a specific game by its ID
     const [rows] = await this.database.query(
       `SELECT game.*, genre.label as genre_label from ${this.table} LEFT JOIN genre ON game.genre_id = genre.id
-    WHERE game.id = ?
+    WHERE game.id = ? AND game.user_id = ?
   `,
-      [id]
+      [id, userId]
     );
 
     // Return the first row of the result, which represents the game
     return rows[0];
   }
 
-  async readAll(searchTerm) {
-    let query = `SELECT game.*, genre.label as genre_label FROM ${this.table} LEFT JOIN genre ON game.genre_id = genre.id`;
-    let params = [];
+  async readAll(userId, searchTerm) {
+    let query = `SELECT game.*, genre.label as genre_label FROM ${this.table} LEFT JOIN genre ON game.genre_id = genre.id WHERE game.user_id = ?`;
+    const params = [userId];
 
     if (searchTerm) {
       query += ` WHERE title LIKE ?`;
-      params = [`%${searchTerm}%`];
+      params.push = [`%${searchTerm}%`];
     } else {
       query += " LIMIT 8";
     }
@@ -56,11 +56,19 @@ class GameManager extends AbstractManager {
   // TODO: Implement the update operation to modify an existing game
 
   // async update(game) {
-  async update(id, game) {
+  async update(userId, id, game) {
     // Execute the SQL INSERT query to add a new game to the "game" table
     const [result] = await this.database.query(
-      `UPDATE ${this.table} SET title = ?, image = ?, year = ?, console = ?, genre_id= ? WHERE id = ?`,
-      [game.title, game.image, game.year, game.console, game.genre_id, id]
+      `UPDATE ${this.table} SET title = ?, image = ?, year = ?, console = ?, genre_id= ? WHERE id = ? AND user_id = ?`,
+      [
+        game.title,
+        game.image,
+        game.year,
+        game.console,
+        game.genre_id,
+        id,
+        userId,
+      ]
     );
 
     // Return the ID of the newly inserted game
@@ -70,12 +78,12 @@ class GameManager extends AbstractManager {
   // }
 
   // The D of CRUD - Delete operation
-  async delete(id) {
+  async delete(userId, id) {
     try {
       // Execute the SQL DELETE query to delete a specific game by its ID
       const result = await this.database.query(
-        `DELETE FROM ${this.table} WHERE id = ?`,
-        [id]
+        `DELETE FROM ${this.table} WHERE id = ? AND user_id = ?`,
+        [id, userId]
       );
 
       // Check the affectedRows property to verify if the deletion was successful
